pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gtm-ga4-sync"
version = "0.1.0"
description = "Declare dataLayer events in YAML, provision GTM triggers/variables/tags + GA4 custom dimensions in one command."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Kenneth J. Buchanan", email = "kennethbuchanan23@gmail.com" },
]
keywords = [
    "gtm",
    "google-tag-manager",
    "ga4",
    "google-analytics",
    "analytics",
    "datalayer",
    "event-tracking",
    "consent-mode",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "google-auth>=2.28",
    "google-auth-oauthlib>=1.2",
    "google-api-python-client>=2.120",
    "google-analytics-admin>=0.22",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "ruff>=0.4",
    "mypy>=1.8",
]

[project.urls]
Homepage = "https://github.com/kb223/gtm-ga4-sync"
Author = "https://kennethjbuchanan.com"

[project.scripts]
gtm-ga4-sync = "gtm_ga4_sync.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
